Header Attributes Difference Check Bug (#7)
**Describe your changes**
Removed the `Project` enum case. The asymmetrical `?? "nil (Project)"` on the second attributes was removed. We assume nil means it is a `Project` type. As a side note, if a header has the same attribute that is not one of public, private or nil (project) then we do not flag it.

**Testing performed**
`HeadersComparatorTests` has been updated. And the integration tests have been updated.