Fix SettingsComparator to handle missing configurations (#35)
**Describe your changes**
For some projects, `SettingsComparator` can return enigmatic error messages "ERROR: Configuration not found". We didn't expect to the targets to contain an empty `XCConfigurationList` but it looks, it's a valid scenario. 
Update the logic so an empty list is handled in the same way as missing `XCConfigurationList`. When one of the targets has a valid configuration object, but the second does not have configuration or has an empty configuration list, "only in..." error will be triggered.

**Testing performed**
CI

Signed-off-by: Marcin Iwanicki <miwanicki1@bloomberg.net>